requirements.txt
streamlit==1.42.0
psycopg2-binary==2.9.9
pandas==2.2.3
plotly==5.24.1
python-dotenv==1.0.1
sqlalchemy==2.0.36
openpyxl==3.1.5
xlsxwriter==3.2.0
Pillow==11.0.0
schedule==1.2.2
pytz==2024.2
babel==2.16.0
streamlit-aggrid==1.0.5
st-aggrid==0.3.4.post3
numpy==2.1.3
